feat(reps): use front side to track reps
